% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auk-package.r
\docType{package}
\name{auk-package}
\alias{auk}
\alias{auk-package}
\title{\code{auk}: eBird Data Extraction and Processing in R}
\description{
Tools for extracting and processing eBird data from the eBird Basic Dataset
(EBD).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cornelllabofornithology.github.io/auk/}
  \item Report bugs at \url{https://github.com/CornellLabofOrnithology/auk/issues}
}

}
\author{
\strong{Maintainer}: Matthew Strimas-Mackey \email{mes335@cornell.edu} (\href{https://orcid.org/0000-0001-8929-7776}{ORCID})

Authors:
\itemize{
  \item Eliot Miller
  \item Wesley Hochachka
}

Other contributors:
\itemize{
  \item Cornell Lab of Ornithology [copyright holder]
}

}
\keyword{internal}
