% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auk-package.r
\docType{package}
\name{auk}
\alias{auk}
\alias{auk-package}
\title{\code{auk}: eBird Data Extraction with AWK}
\description{
Tools for extracting and processing eBird data from the eBird Basic Dataset
(EBD).
}
