% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auk-ebd.r
\name{auk_ebd}
\alias{auk_ebd}
\title{Reference to EBD file}
\usage{
auk_ebd(file, file_sampling, sep = "\\t")
}
\arguments{
\item{file}{character; input file.}

\item{file_sampling}{character; optional input sampling event data file,
required if you intend to zero-fill the data to produce a presence absence
data set. The sampling file consists of just effort information for every
eBird checklist. Any species not appearing in the EBD for a given checklist
is implicitly considered to have a count of 0. This file should be
downloaded at the same time as the EBD to ensure they are in sync.}

\item{sep}{character; the input field seperator, the EBD is tab separated so
this should generally not be modified. Must only be a single character and
space delimited is not allowed since spaces appear in many of the fields.}
}
\value{
An \code{ebd} object storing the file reference and the desired filters
once created with other package functions.
}
\description{
Create a reference to an eBird Basic Dataset (EBD) file in preparation for
filtering using AWK.
}
\details{
The EBD can be downloaded as a tab-separated text file from the
\href{http://ebird.org/ebird/data/download}{eBird website} after submitting a
request for access. As of February 2017, this file is nearly 150 GB making
it challenging to work with. If you're only interested in a single species
or a small region it is possible to submit a custom download request. This
approach is suggested to speed up processing time.
}
\examples{
# example data
f <- system.file("extdata/ebd-sample.txt", package = "auk")
auk_ebd(f)
}
