% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auk-filter.r
\name{auk_filter}
\alias{auk_filter}
\title{Filter the EBD using AWK}
\usage{
auk_filter(x, file, awk_file, sep, execute, overwrite)
}
\arguments{
\item{x}{\code{ebd} object; reference to EBD file created by \code{\link[=auk_ebd]{auk_ebd()}}.}

\item{file}{character; output file.}

\item{awk_file}{character; output file to optionally save the awk script to.}

\item{sep}{character; the input field seperator, the EBD is tab separated by
default. Must only be a single character and space delimited is not allowed
since spaces appear in many of the fields.}

\item{execute}{logical; whether to execute the awk script, or output it to a
file for manual execution. If this flag is \code{FALSE}, \code{awk_file} must be
provided.}

\item{overwrite}{logical; overwrite output file if it already exists}
}
\value{
If AWK ran without errors, the output filename is returned,
however, if an error was encountered the exit code is returned. If \code{execute = FALSE}, then the path to the AWK script is returned rather than the path
to the output file.
}
\description{
Convert the filters defined in an \code{ebd} object into an AWK script and run
this script to produce a filtered eBird Reference Dataset (ERD). The initial
creation of the \code{ebd} object should be done with \code{\link[=auk_ebd]{auk_ebd()}} and filters can
be defined using the various other functions in this package, e.g.
\code{\link[=auk_species]{auk_species()}} or \code{\link[=auk_country]{auk_country()}}. Note that this function typically takes
at least a couple hours to run.
}
\details{
The AWK script can be saved for future reference by providing an output
filename to \code{awk_file}. The default behvaiour of this function is to generate
and run the AWK script, however, by setting \code{execute = FALSE} the AWK script
will be generated but not run. In this case, \code{file} is ignored and \code{awk_file}
must be specified.

Calling this function requires that the command line utility AWK is
installed. Linux and Mac machines should have AWK by default, Windows users
will likely need to install \href{https://www.cygwin.com}{Cygwin}.
}
\examples{
# define filters
filters <- system.file("extdata/ebd-sample.txt", package="auk") \%>\%
  auk_ebd() \%>\%
  auk_species(species = c("Gray Jay", "Blue Jay")) \%>\%
  auk_country(country = c("US", "Canada")) \%>\%
  auk_extent(extent = c(-100, 37, -80, 52)) \%>\%
  auk_date(date = c("2012-01-01", "2012-12-31")) \%>\%
  auk_time(time = c("06:00", "09:00")) \%>\%
  auk_duration(duration = c(0, 60)) \%>\%
  auk_complete()
# temp output file
out_file <- tempfile()
auk_filter(filters, file = out_file)
# clean
unlink(out_file)
}
