% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auk-get-awk-path.R
\name{auk_get_awk_path}
\alias{auk_get_awk_path}
\title{OS specific path to AWK executable}
\usage{
auk_get_awk_path()
}
\value{
Path to AWK or \code{NA} if AWK wasn't found.
}
\description{
Return the OS specific path to AWK (e.g. \code{"C:/cygwin64/bin/gawk.exe"} or
\code{"/usr/bin/awk"}), or highlights if it's not installed. To manually set the
path to AWK, set the \code{AWK_PATH} environment variable in your \code{.Renviron}
file, which can be accomplished with the helper function
\code{auk_set_awk_path(path)}.
}
\examples{
auk_get_awk_path()
}
\seealso{
Other paths: 
\code{\link{auk_get_ebd_path}()},
\code{\link{auk_set_awk_path}()},
\code{\link{auk_set_ebd_path}()}
}
\concept{paths}
