% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auk-get-ebd-path.R
\name{auk_get_ebd_path}
\alias{auk_get_ebd_path}
\title{Return EBD data path}
\usage{
auk_get_ebd_path()
}
\value{
The path stored in the \code{EBD_PATH} environment variable.
}
\description{
Returns the environment variable \code{EBD_PATH}, which users are encouraged to
set to the directory that stores the eBird Basic Dataset (EBD) text files.
}
\examples{
auk_get_ebd_path()
}
\seealso{
Other paths: 
\code{\link{auk_get_awk_path}()},
\code{\link{auk_set_awk_path}()},
\code{\link{auk_set_ebd_path}()}
}
\concept{paths}
