% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auk-installed.r
\name{auk_installed}
\alias{auk_installed}
\title{Check for a valid AWK install}
\usage{
auk_installed()
}
\value{
A logical indicating whether awk was found or not.
}
\description{
Check for a valid AWK install
}
\examples{
auk_installed()
}
