% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auk-read.r
\name{auk_read}
\alias{auk_read}
\title{Read an EBD file}
\usage{
auk_read(file)
}
\arguments{
\item{file}{character; EBD file to read.}
}
\value{
A \code{data.frame}.
}
\description{
A thin wrapper for \link{read.delim} that reads EBD files and cleans on the
variable names.
}
\examples{
ebd <- system.file("extdata/ebd-sample.txt", package="auk") \%>\%
  auk_read()
}
