% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auk-set-awk-path.r
\name{auk_set_awk_path}
\alias{auk_set_awk_path}
\title{Set a custom path to AWK executable}
\usage{
auk_set_awk_path(path)
}
\arguments{
\item{path}{character; path to the AWK executable on your system, e.g.
\code{"C:/cygwin64/bin/gawk.exe"} or \code{"/usr/bin/awk"}.}
}
\value{
Opens .Renviron for editing and returns the AWK path invisibly.
}
\description{
If AWK has been installed in a non-standard location, the environment
variable \code{AWK_PATH} must be set to specify the location of the executable.
This function is a helper for editing the .Renviron file to add this path.
\strong{Most users should NOT set \code{AWK_PATH}, only do so if you have installed AWK
in non-standard location and \code{auk} cannot find it.}
}
\examples{
\dontrun{
auk_set_awk_path("/usr/bin/awk")
}
}
\seealso{
Other paths: \code{\link{auk_get_awk_path}},
  \code{\link{auk_get_ebd_path}},
  \code{\link{auk_set_ebd_path}}
}
\concept{paths}
