% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auk-version-date.r
\name{auk_version_date}
\alias{auk_version_date}
\title{Dates of EBD and Taxonomy in package version}
\usage{
auk_version_date()
}
\value{
A date vector with the EBD date and taxonomy date that this version
of the package corresponds to.
}
\description{
This package depends on the version of the EBD and on the eBird taxonomy. Use
this function to determine the version dates for which the package is
suitable.
}
\examples{
auk_version_date()
}
