% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ebird_states}
\alias{ebird_states}
\title{eBird States}
\format{A data frame with four variables and 3,145 rows:
\itemize{
\item \code{country}: short form of English country name.
\item \code{country_code}: 2-letter ISO country code.
\item \code{state}: state name.
\item \code{state_code}: 4 to 6 character state code.
}}
\usage{
ebird_states
}
\description{
A data frame of state codes used by eBird. These codes are 4 to 6 characters,
consisting of two parts, the 2-letter ISO country code and a 1-3 character
state code, separated by a dash. For example, \code{"US-NY"} corresponds to New
York State in the United States. These state codes are required to filter by
state using \code{\link[=auk_state]{auk_state()}}.
}
\details{
Note that some countries are not broken into states in eBird and therefore do
not appear in this data frame.
}
\seealso{
Other data: \code{\link{ebird_taxonomy}}
}
\concept{data}
\keyword{datasets}
