% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{read_ebd}
\alias{read_ebd}
\title{Read an EBD file}
\usage{
read_ebd(file, sep = "\\t", setclass = c("data.frame", "tbl", "data.table"))
}
\arguments{
\item{file}{character; EBD file to read.}

\item{sep}{character; single character used to separate fields within a row.}

\item{setclass}{\code{data.frame}, \code{tbl}, or \code{data.table}; optionally set
additional classes to set on the output data. All return objects are
data frames, but may additionally be \code{tbl} (for use with \code{dplyr}
and the tidyverse) or \code{data.table} (for use with \code{data.table}).}
}
\value{
A \code{data.frame} unless \code{setclass} is used in which case the data frame
will have additonal classes \code{tbl} or \code{data.table}.
}
\description{
Read an eBird Basic Dataset file using \link[=fread]{data.table},
\link[=read_delim]{readr}, or \link{read.delim} depending on which packages are
installed. This function chooses read options that are suitable for EBD files
and cleans up variable names to be in \code{snake_case}.
}
\examples{
ebd <- system.file("extdata/ebd-sample.txt", package="auk") \%>\%
  read_ebd()
# optionally return a tibble for use with tidyverse
ebd_tbl <- system.file("extdata/ebd-sample.txt", package="auk") \%>\%
  read_ebd(setclass = "tbl")
}
