% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-zerofill.r
\name{read_zerofill}
\alias{read_zerofill}
\alias{read_zerofill.character}
\alias{read_zerofill.auk_ebd}
\title{Read and zero-fill an EBD file}
\usage{
read_zerofill(x, ...)

\method{read_zerofill}{character}(x, sampling_events, sep = "\\t",
  setclass = c("tbl", "data.frame", "data.table"), ...)

\method{read_zerofill}{auk_ebd}(x, sep = "\\t", setclass = c("tbl",
  "data.frame", "data.table"), ...)
}
\arguments{
\item{x}{filename or \code{auk_ebd} object with associtated output
files as created by \code{\link[=auk_filter]{auk_filter()}}. If a filename is provided, it must
point to the EBD and the \code{sampling_events} argument must point to the
sampling event data file.}

\item{...}{additional arguments passed to methods.}

\item{sampling_events}{character; filename for the sampling event data.}

\item{sep}{character; single character used to separate fields within a row.}

\item{setclass}{\code{tbl}, \code{data.frame}, or \code{data.table}; optionally set
additional classes to set on the output data. All return objects are
data frames, but may additionally be \code{tbl} (for use with \code{dplyr}
and the tidyverse) or \code{data.table} (for use with \code{data.table}). The default
is to return a tibble.}
}
\value{
A \code{data.frame} with additional class \code{tbl} unless \code{setclass} is used,
in which case a standard \code{data.frame} or \code{data.table} can be returned.
}
\description{
Read an eBird Basic Dataset file, and associated sampling event data file, to
produce a zero-filled, presence-absence dataset. The EBD contains bird
sightings and the sampling event data is a set of all checklists, they can be
combined to infer absence data by assuming any species not reported on a
checklist was had a count of zero.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Filename of EBD.

\item \code{auk_ebd}: \code{auk_ebd} object output from \code{\link[=auk_filter]{auk_filter()}}. Must
have had a sampling event data file set in the original call to
\code{\link[=auk_ebd]{auk_ebd()}}.
}}

\examples{
# read and zero-fill the sampling data
f_ebd <- system.file("extdata/zerofill-ex_ebd.txt", package = "auk")
f_smpl <- system.file("extdata/zerofill-ex_sampling.txt", package = "auk")
x <- read_zerofill(x = f_ebd, sampling_events = f_smpl)
}
