% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{valid_protocols}
\alias{valid_protocols}
\title{Valid Protocols}
\format{A vector with 42 elements.}
\usage{
valid_protocols
}
\description{
A vector of valid protocol names, e.g. "Traveling", "Stationary", etc.
}
\keyword{datasets}
