% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auk-ebd.r
\name{auk_ebd}
\alias{auk_ebd}
\title{Reference to EBD file}
\usage{
auk_ebd(file, sep = "\\t")
}
\arguments{
\item{file}{character; input file.}

\item{sep}{character; the input field seperator, the EBD is tab separated
by default. Must only be a single character and space delimited is not
allowed since spaces appear in many of the fields.}
}
\value{
An \code{ebd} object storing the file reference and the desired filters
once created with other package functions.
}
\description{
Create a reference to an eBird Basic Dataset (EBD) file in preparation for
filtering using AWK.
}
\examples{
# example data
f <- system.file("extdata/ebd-sample.txt", package="auk")
auk_ebd(f)
}
