% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auk-getpath.r
\name{auk_getpath}
\alias{auk_getpath}
\title{OS specific path to AWK}
\usage{
auk_getpath()
}
\value{
Path to AWK or \code{NA} if AWK wasn't found.
}
\description{
Return the OS specific path to AWK, or highlights if it's not installed. To
manually set the path to AWK, set the \code{AWK_PATH} environment variable in your
\code{.Renviron} file.
}
\examples{
auk_getpath()
}
