% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auk-unique.r
\name{auk_unique}
\alias{auk_unique}
\title{Remove duplicate group checklists}
\usage{
auk_unique(x, group_id = "group_identifier",
  checklist_id = "sampling_event_identifier")
}
\arguments{
\item{x}{data.frame; the EBD data frame, typically as imported by
\code{\link[=read_ebd]{read_ebd()}}.}

\item{group_id}{character; the name of the group ID column.}

\item{checklist_id}{character; the name of the checklist ID column, each
checklist within a group will get a unique value for this field. The record
with the lowest \code{checklist_id} will be picked as the unique record within
each group.}
}
\value{
A data.frame with unique observations.
}
\description{
eBird checklists can be shared among a group of multiple observers, in which
case observations will be duplicated in the database. This functions removes
these duplicates from the eBird Basic Dataset (EBD) creating a set of unique
bird observations. This function is called automatically by \code{\link[=read_ebd]{read_ebd()}}.
}
\details{
This function chooses the checklist within in each that has the
lowest value for the field specified by \code{checklist_id}. A new column is
also created, \code{checklist_id}, whose value is the taken from the field
specified in the \code{checklist_id} parameter for non-group checklists and from
the field specified by the \code{group_id} parameter for grouped checklists.
}
\examples{
# read in an ebd file and don't automatically remove duplicates
ebd <- system.file("extdata/ebd-sample.txt", package = "auk") \%>\%
  read_ebd(unique = FALSE)
# remove duplicates
ebd_unique <- auk_unique(ebd)
nrow(ebd)
nrow(ebd_unique)
}
