% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auk-version-date.r
\name{auk_version_date}
\alias{auk_version_date}
\title{Dates of eBird Basic Dataset and taxonomy in package version}
\usage{
auk_version_date()
}
\value{
A date vector with the eBird data date and taxonomy date that this
version of the package corresponds to.
}
\description{
This package depends on the version of the EBD and on the eBird taxonomy. Use
this function to determine the version dates for which the package is
suitable. The EBD is update quarterly (March 15, June 15, September 15,
December 15), while the taxonomy is updated annually in September. To ensure
proper functioning, always use the latest version of the auk package and the
EBD.
}
\examples{
auk_version_date()
}
\seealso{
Other helpers: \code{\link{auk_getpath}},
  \code{\link{ebird_species}}
}
\concept{helpers}
