% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ebd.r
\name{read_ebd}
\alias{read_ebd}
\title{Read an EBD file}
\usage{
read_ebd(x, sep, unique, setclass)
}
\arguments{
\item{x}{filename of EBD file or \code{auk_ebd} object with associtated output
files as created by \code{\link[=auk_filter]{auk_filter()}}.}

\item{sep}{character; single character used to separate fields within a row.}

\item{unique}{logical; should duplicate grouped checklists be removed. If
\code{unique = TRUE}, \code{\link[=auk_unique]{auk_unique()}} is called on the EBD before returning.}

\item{setclass}{\code{tbl}, \code{data.frame}, or \code{data.table}; optionally set
additional classes to set on the output data. All return objects are
data frames, but may additionally be \code{tbl} (for use with \code{dplyr}
and the tidyverse) or \code{data.table} (for use with \code{data.table}). The default
is to return a tibble.}
}
\value{
A \code{data.frame} with additional class \code{tbl} unless \code{setclass} is used,
in which case a standard \code{data.frame} or \code{data.table} can be returned.
}
\description{
Read an eBird Basic Dataset file using \code{\link[data.table:fread]{data.table::fread()}},
\code{\link[readr:read_delim]{readr::read_delim()}}, or \link{read.delim} depending on which packages are
installed. \strong{Note that this function typically takes at least a couple hours
to run.}
}
\details{
This functions performs the following processing steps:
\itemize{
\item Data types for columns are manually set based on column names used in the
February 2017 EBD. If variables are added or names are changed in later
releases, any new variables will have data types inferred by the import
function used.
\item Variables names are converted to \code{snake_case}.
\item Duplicate observations resulting from group checklists are removed using
\code{\link[=auk_unique]{auk_unique()}}, unless \code{unique = FALSE}.
}
}
\examples{
ebd <- system.file("extdata/ebd-sample.txt", package = "auk") \%>\%
  read_ebd()
# optionally return a plain data.frame
ebd_df <- system.file("extdata/ebd-sample.txt", package = "auk") \%>\%
  read_ebd(setclass = "data.frame")
}
